+++ makefile/lint.py
@@ -17,6 +17,9 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 -----------------------------------------------------------------------
 """
 
+# TODO:
+# Detect if function should be static
+
 from pathlib import Path
 
 import makefile.globals
@@ -82,7 +85,7 @@ def lint():
                     print(f"\033[92mTODO\033[0m       {file}")
                     print(f"    \033[90m{clean_line}\033[0m")
 
-                elif "LOG_CALL" in clean_line:
+                elif "LOG_CALL" in clean_line and file.parts[-1] != "kernel.h":
                     print(f"\033[92mLOG_CALL\033[0m   {file}")
 
                 if clean_line.startswith("#include"):